+++ CHANGES.md
@@ -13,6 +13,7 @@
 * Update typing-extensions 4.14.0 (59d2c20) to 4.15.0 (2638b86)
 * Update urllib3 2.5.0 (98a60f9) to 2.6.2 (83f8643)
 * Update zoneinfo 2025a to 2025c (fallback)
+* Change improve Proper and Series name search for Newznab provider type
 
 
 [develop]